+++ zarcel-core/src/main/java/com/zing/zalo/data/serialization/SerializedByteBufferInput.java
@@ -1,5 +1,6 @@
 package com.zing.zalo.data.serialization;
 
+import java.io.UnsupportedEncodingException;
 import java.nio.ByteBuffer;
 
 /**
@@ -8,12 +9,10 @@ import java.nio.ByteBuffer;
 public class SerializedByteBufferInput implements SerializedInput {
 
     private ByteBuffer in;
-    private int position;
     private int length;
 
     public SerializedByteBufferInput(byte data[]) {
         in = ByteBuffer.wrap(data);
-        position = 0;
         length = data.length;
     }
 
@@ -31,9 +30,7 @@ public class SerializedByteBufferInput implements SerializedInput {
     @Override
     public int readByte() {
         try {
-            int ret = (in.get() & 0xFF);
-            ++position;
-            return ret;
+            return (int) readByteHeader();
         } catch (Exception e) {
             throw new RuntimeException("read byte error", e);
         }
@@ -42,12 +39,7 @@ public class SerializedByteBufferInput implements SerializedInput {
     @Override
     public int readInt16() {
         try {
-            int i = 0;
-            for (int j = 0; j < 2; j++) {
-                i |= ((in.get() & 0xFF) << (j * 8));
-                position++;
-            }
-            return i;
+            return (int) readShortHeader();
         } catch (Exception e) {
             throw new RuntimeException("readInt16: read byte error", e);
         }
@@ -56,12 +48,7 @@ public class SerializedByteBufferInput implements SerializedInput {
     @Override
     public int readInt32() {
         try {
-            int i = 0;
-            for (int j = 0; j < 4; j++) {
-                i |= ((in.get() & 0xFF) << (j * 8));
-                position++;
-            }
-            return i;
+            return readIntHeader();
         } catch (Exception e) {
             throw new RuntimeException("read byte error", e);
         }
@@ -70,19 +57,13 @@ public class SerializedByteBufferInput implements SerializedInput {
 
     @Override
     public boolean readBool() {
-        return readByte() == 1;
+        return readBoolHeader();
     }
 
     @Override
     public long readInt64() {
         try {
-            long i = 0;
-            for (int j = 0; j < 8; j++) {
-                long tmp = ((long) in.get() & 0xFF) << (j * 8);
-                i |= tmp;
-                position++;
-            }
-            return i;
+            return readLongHeader();
         } catch (Exception e) {
             throw new RuntimeException("read byte error", e);
         }
@@ -91,14 +72,7 @@ public class SerializedByteBufferInput implements SerializedInput {
     @Override
     public String readString() {
         try {
-            int l = readInt32();
-            if (l < 0) return null;
-            if (l == 0) return "";
-            if (l > length) throw new RuntimeException("read byte length error");
-            byte[] b = new byte[l];
-            in.get(b);
-            position += l;
-            return new String(b, "UTF-8");
+            return readStringHeader();
         } catch (Exception e) {
             throw new RuntimeException("read string error", e);
         }
@@ -107,12 +81,7 @@ public class SerializedByteBufferInput implements SerializedInput {
     @Override
     public byte[] readByteArray() {
         try {
-            int l = readInt32();
-            if (l > length) throw new RuntimeException("read byte length error");
-            byte[] b = new byte[l];
-            in.get(b);
-            position += l;
-            return b;
+            return readBytesHeader();
         } catch (Exception e) {
             throw new RuntimeException("read byte error", e);
         }
@@ -120,7 +89,7 @@ public class SerializedByteBufferInput implements SerializedInput {
 
     @Override
     public double readDouble() {
-        return Double.longBitsToDouble(readInt64());
+        return readDoubleHeader();
     }
 
     @Override
@@ -130,18 +99,137 @@ public class SerializedByteBufferInput implements SerializedInput {
         }
         if (in != null) {
             try {
-                byte[] tmp = new byte[count];
-                in.get(tmp);
-                position += count;
+                in.position(in.position() + count);
             } catch (Exception e) {
                 throw new RuntimeException("skip error", e);
             }
         }
     }
 
+    /**
+     * @param expectedType expected type.
+     * @return subtype when expectedType == obtainedType
+     */
+    public byte readHeader(byte expectedType) {
+        byte obtainedType = in.get();
+        if (ZarcelDefs.getHeaderType(obtainedType) == expectedType) {
+            return ZarcelDefs.getHeaderSubType(obtainedType);
+        }
+        throw new RuntimeException("Expected type: " + ZarcelDefs.getTypeName(expectedType) + ", but serialization type: " + ZarcelDefs.getTypeName(obtainedType));
+    }
+
+    private boolean readBoolHeader() {
+        byte subType = readHeader(ZarcelDefs.TYPE_BOOLEAN);
+        if (subType == ZarcelDefs.SUBTYPE_ZERO)
+            return false;
+        return true;
+    }
+
+    private byte readByteHeader() {
+        byte subtype = readHeader(ZarcelDefs.TYPE_BYTE);
+        if (subtype == ZarcelDefs.SUBTYPE_ZERO)
+            return 0;
+        return in.get();
+    }
+
+    private short readShortHeader() {
+        byte subtype = readHeader(ZarcelDefs.TYPE_SHORT);
+        switch (subtype) {
+            case ZarcelDefs.SUBTYPE_ZERO:
+                return 0;
+            case ZarcelDefs.SUBTYPE_BYTE:
+                return (short) (in.get() & 0x00FF);
+            default:
+                return in.getShort();
+        }
+    }
+
+    private int readIntHeader() {
+        byte subtype = readHeader(ZarcelDefs.TYPE_INT);
+        switch (subtype) {
+            case ZarcelDefs.SUBTYPE_ZERO:
+                return 0;
+            case ZarcelDefs.SUBTYPE_BYTE:
+                return (in.get() & 0x000000FF);
+            case ZarcelDefs.SUBTYPE_SHORT:
+                return (in.getShort() & 0x0000FFFF);
+            default:
+                return in.getInt();
+        }
+    }
+
+    private long readLongHeader() {
+        byte subtype = readHeader(ZarcelDefs.TYPE_LONG);
+        switch (subtype) {
+            case ZarcelDefs.SUBTYPE_ZERO:
+                return 0;
+            case ZarcelDefs.SUBTYPE_BYTE:
+                return (in.get() & 0x00000000000000FFL);
+            case ZarcelDefs.SUBTYPE_SHORT:
+                return (in.getShort() & 0x000000000000FFFFL);
+            case ZarcelDefs.SUBTYPE_INT:
+                return (in.getInt() & 0x00000000FFFFFFFFL);
+            default:
+                return in.getLong();
+        }
+    }
+
+    private double readDoubleHeader() {
+        byte subtype = readHeader(ZarcelDefs.TYPE_DOUBLE);
+        long result;
+        switch (subtype) {
+            case ZarcelDefs.SUBTYPE_ZERO:
+                result = 0;
+                break;
+            case ZarcelDefs.SUBTYPE_BYTE:
+                result = (in.get() & 0x00000000000000FFL);
+                break;
+            case ZarcelDefs.SUBTYPE_SHORT:
+                result = (in.getShort() & 0x000000000000FFFFL);
+                break;
+            case ZarcelDefs.SUBTYPE_INT:
+                result = (in.getInt() & 0x00000000FFFFFFFFL);
+                break;
+            default:
+                result = in.getLong();
+                break;
+        }
+        return Double.longBitsToDouble(result);
+    }
+
+    private byte[] readBytesHeader() {
+        byte subtype = readHeader(ZarcelDefs.TYPE_BYTE_ARRAY);
+        switch (subtype) {
+            case ZarcelDefs.SUBTYPE_ZERO:
+                return new byte[0];
+            default:
+                int length = readIntHeader();
+                if (length > this.length) throw new RuntimeException("read byte length error");
+                byte[] result = new byte[length];
+                in.get(result, 0, length);
+                return result;
+        }
+    }
+
+    private String readStringHeader() throws UnsupportedEncodingException {
+        byte subtype = readHeader(ZarcelDefs.TYPE_STRING);
+        switch (subtype) {
+            case ZarcelDefs.SUBTYPE_ZERO:
+                return "";
+            case ZarcelDefs.SUBTYPE_NULL:
+                return null;
+            default:
+                int length = readIntHeader();
+                if (length > this.length) throw new RuntimeException("read byte length error");
+                byte[] result = new byte[length];
+                in.get(result, 0, length);
+                return new String(result, "UTF-8");
+        }
+    }
+
     @Override
     public int getPosition() {
-        return position;
+        return in.position();
     }
 
     @Override